requirements.txt
supabase==2.3.4
python-dotenv==1.0.0
reportlab==4.0.7
Pillow==10.1.0
